apps/scraper/requirements.txt
# AtlasReg Scraper - Python Dependencies

# Web Scraping
scrapy==2.11.0
playwright==1.40.0
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1

# Storage
minio==7.2.0
boto3==1.29.7

# Database
sqlalchemy>=1.4.28,<2.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-celery==3.4.1

# Celery
celery==5.3.4
redis==5.0.1

# AI/ML
torch==2.1.1
transformers==4.35.2
spacy==3.7.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105
pytesseract==0.3.10
pdf2image==1.16.3

# NLP Portuguese
nltk==3.8.1

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# Monitoring
flower==2.0.1
prometheus-client==0.19.0
